requirements.txt
aiogram==3.25.0
aiosqlite==0.22.1
pandas==3.0.1
matplotlib==3.10.8
seaborn==0.13.2
python-dotenv==1.2.1
reportlab==4.4.10
numpy==2.4.2
xlsxwriter==3.2.9
flask==3.1.3
flask-cors==6.0.2
gunicorn==25.1.0